requirements.txt
python-telegram-bot==13.15
telethon==1.33.1
requests